requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
exceptiongroup==1.1.1
Flask==2.3.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kaggle==1.5.13
MarkupSafe==2.1.2
packaging==23.1
pluggy==1.0.0
pymongo==4.3.3
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
requests==2.31.0
six==1.16.0
text-unidecode==1.3
tomli==2.0.1
tqdm==4.65.0
urllib3==1.26.6
Werkzeug==2.3.4
zipp==3.15.0
